Fix crash on data products with lower than 10 count
In this change, we check whether the number of features
is not lower than the count. If it is lower, we got out
of bounds errors since it wanted to take a slice of
the feature that it didn't have.

Therefore we now set the count equal to the amount of
features if the count is higher.